Updating KnowledgeBase page with Tanstack
